# openapi.yaml
openapi: "3.1.0"
info:
  title: MTS IaaS Cloud
  description: |
    Управление виртуальными машинами, проектами, пользователями и сетями
    на платформе MTS IaaS Cloud. Этот файл — единый источник истины (source of truth)
    для бэкенда и фронтенда.
  version: "1.0.0"
  contact:
    name: MTS IaaS Team
servers:
  - url: http://localhost:8000
    description: Local dev
  - url: http://localhost:4010
    description: Prism mock server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Uuid:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"

    # ── Ошибки ──────────────────────────────────────
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
      example:
        detail: "Resource not found"
    ValidationError:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string

    # ── Auth / User ─────────────────────────────────
...
    ProjectStatus:
      type: string
      enum:
        - CREATING
        - ACTIVE
        - INACTIVE
        - DELETING
        - ERROR
    ProjectCreateRequest:
      type: object
      required: [name, cpu_quota, ram_quota, ssd_quota]
      properties:
        name:
          type: string
          minLength: 1
          example: "production-project"
        cpu_quota:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
          example: 8
        ram_quota:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
          example: 16384
        ssd_quota:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
          example: 100
    ProjectUpdateRequest:
      type: object
      required: [cpu_quota, ram_quota, ssd_quota]
      properties:
        cpu_quota:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
        ram_quota:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
        ssd_quota:
          type: integer
          format: int32
          minimum: 1
          maximum: 2147483647
    ProjectResponse:
      type: object
      required: [id, name, cpu_quota, ram_quota, ssd_quota, cpu_used, ram_used, ssd_used, is_allocated, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        owner_id:
          type: string
          format: uuid
          nullable: true
        cpu_quota:
          type: integer
        ram_quota:
          type: integer
        ssd_quota:
          type: integer
        cpu_used:
          type: integer
        ram_used:
          type: integer
        ssd_used:
          type: integer
        is_allocated:
          type: boolean
        status:
          $ref: "#/components/schemas/ProjectStatus"
      example:
        id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
        name: "production-project"
        owner_id: "550e8400-e29b-41d4-a716-446655440000"
        cpu_quota: 8
        ram_quota: 16384
        ssd_quota: 100
        cpu_used: 2
        ram_used: 4096
        ssd_used: 50
        is_allocated: true
        status: "ACTIVE"
    ProjectActionResponse:
      type: object
      required: [job_id, status]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
      example:
        job_id: "c3d4e5f6-a7b8-9012-cdef-123456789012"
        status: "pending"

    # ── Admin schemas ───────────────────────────────
...
    AdminVMResponse:
      type: object
      required: [id, name, project_id, cpu, ram, ssd, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        project_id:
          type: string
          format: uuid
        cpu:
          type: integer
        ram:
          type: integer
        ssd:
          type: integer
        status:
          type: string
        docker_container_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    AdminServersListResponse:
      type: object
      required: [servers]
      properties:
        servers:
          type: array
          items:
            $ref: "#/components/schemas/AdminVMResponse"
...
    FileLoaderResponse:
      type: object
      required: [project_root, status, mode]
      properties:
        project_root:
          type: string
        status:
          type: string
          enum: [uploaded]
        mode:
          type: string
          enum: [archive, file]
        filename:
          type: string
          nullable: true
    AskResponse:
      type: object
      required: [response, docker_created]
      properties:
        response:
          type: string
        specs:
          type: object
          nullable: true
        docker_created:
          type: boolean
        project_root_used:
          type: string
          nullable: true
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: "Resource not found"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: "Not authenticated"
    Forbidden:
      description: Admin only
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: "Admin only"
    BadRequest:
      description: Invalid input (malformed HTTP, body parse error, etc.)
      content:
        text/plain:
          schema:
            type: string
          example: "Invalid HTTP request received."
        text/plain; charset=utf-8:
          schema:
            type: string
          example: "Invalid HTTP request received."
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
paths:
...
  /projects:
    post:
      operationId: createProject
      summary: Создать проект (admin)
      tags: [projects]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreateRequest"
      responses:
        "201":
          description: Проект создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"

    get:
      operationId: listProjects
      summary: Список проектов (admin)
      tags: [projects]
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Фильтр по имени проекта (ILIKE)
      responses:
        "200":
          description: Список проектов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /projects/{project_id}/start:
    post:
      operationId: startProject
      summary: Запустить проект
      tags: [projects]
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "202":
          description: Job enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectActionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /projects/{project_id}/stop:
    post:
      operationId: stopProject
      summary: Остановить проект
      tags: [projects]
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "202":
          description: Job enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectActionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /projects/{project_id}:
    patch:
      operationId: updateProject
      summary: Обновить квоты проекта
      tags: [projects]
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdateRequest"
      responses:
        "200":
          description: Проект обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    delete:
      operationId: deleteProject
      summary: Удалить проект
      tags: [projects]
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "202":
          description: Job enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectActionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ═══════════════════════════════════════════════════
  #  Admin — Projects
  # ═══════════════════════════════════════════════════
...
  /servers/admin/info/project/{project_id}:
    get:
      operationId: adminListServersByProject
      summary: VM в проекте
      tags: [admin]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Список VM в проекте
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminServersListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ═══════════════════════════════════════════════════
  #  Agent
  # ═══════════════════════════════════════════════════
  /file_loader:
    post:
      operationId: uploadFile
      summary: Загрузить файл/архив проекта
      tags: [agent]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Файл загружен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileLoaderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          description: Файл слишком большой
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /ask:
    post:
      operationId: askAgent
      summary: Задать вопрос AI-агенту
      tags: [agent]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: []
              properties:
                user_message:
                  type: string
                expected_load:
                  type: string
                  enum: [low, medium, high, production]
                  default: medium
                project_root:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Ответ агента
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AskResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          description: Agent execution error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
...